More policies and some bug fixes

+++ app/Http/Controllers/GroupController.php
@@ -41,12 +41,13 @@ class GroupController extends Controller
 
     public function create(Request $request)
     {
-        if ($request->user() === null) {
+        if ($request->user() === null && Group::where('name','=', $request->input('name'))->firstOrFail() !== null) {
             return response()->json(['failure' => 401]);
         }
 
-        // No need for policy ... the user just has to have a authenticated account
-        // e se já houve um com o nome
+        // No need for policy ... the user just only has to have a authenticated account
+        // e se já houve um com o nome .. 
+
 
         $group = new Group();
 
@@ -65,12 +66,12 @@ class GroupController extends Controller
         $group->owners()->save($owner);
         $this->add_topics($request, $group);
 
-        return $group; // TODO : N é isto
+        return response()->json(['The group was created with success' => 200]);
     }
 
     private function add_topics(Request $request, Group $group)
     {   
-        // THIS IS A FUNCTION AND DOES NOT NEEDS POLICY HERE... but in the callee function
+        // THIS IS A FUNCTION AND DOES NOT NEED POLICY HERE... but in the callee function
         if ($request->input('tags') != null) {
 
             $topics = explode(' ', $request->input('tags'));
@@ -95,7 +96,7 @@ class GroupController extends Controller
     public function delete($name)
     {  
         $group = Group::where('name', $name)->first();
-        //$this->authorize('delete', $group); // POLICY
+        $this->authorize('delete', $group); // POLICY
         $group->delete(); 
 
         return response()->json(['The group was deleted with success' => 200]);
@@ -134,6 +135,9 @@ class GroupController extends Controller
 
     private function edit_topics(Request $request, Group $group)
     {
+        //
+        // The policy for this Function is in the function above
+        //
         $group->topics()->delete();
         if ($request->input('tags') != null) {
 
@@ -163,6 +167,8 @@ class GroupController extends Controller
 
         if ($user === null)
             return redirect()->route('home');
+
+        
 
         return view('pages.group_list', ['user' => $user]);
     }
